Biome auto format on commit with Husky & git-format-staged
Set up a git pre-commit hook using Husky to automatically lint staged changes with Biome.

git-format-staged ensures that only the staged changes get checked & formatted.
Unstaged changes in partially-staged files are ignored.

If there are syntax errors or unfixable errors, the commit is blocked. If, however, there
are merely unfixable warnings, the commit is allowed. We could have lint warnings block
commits with Biome's `--error-on-warnings` argument if desired.

Note that the documentation at https://biomejs.dev/recipes/git-hooks/ is incorrect. It
suggests using the file pattern `.` for git-format-staged, but git-format-staged doesn't
support that because it's using Python's fnmatch() function to expand filename patterns.